+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Unreleased
 
-- Fixed integration tests for xDFSNamespaceServerConfiguration.
+- Fixed xDFSNamespaceServerConfiguration by converting LocalHost to ComputerName
+  instead.
 
 ## 3.2.0.0